fix(sales-order): diskon jadi persen bulat, dan beri tahu tim tally
Penyimpanan Sales Order membuang titik dari nilai diskon, sama seperti yang
dilakukannya pada berat dan harga. Untuk berat dan harga itu benar --
JavaScript di form memang memasang titik sebagai pemisah ribuan. Untuk
diskon keliru, karena form sengaja tidak memformat kolom itu, sehingga titik
di sana hanya mungkin berarti koma desimal:

  diketik 2,5%    tersimpan 25%
  diketik 12,75%  tersimpan 1275%

Validasi tidak menangkapnya. Validator melihat 12,75 dan meloloskannya;
perusakannya terjadi sesudah itu, di afterCreate dan afterSave.

Hal ini terjangkau sejak customers.default_discount dibuat bertipe desimal
sementara sales_order_items.discount bertipe bilangan bulat. Keputusan
Project Owner: persen bulat saja. Kedua kolom kini sama, dan aturannya
integer di kedua form. Kalau suatu saat diskon berkoma dibutuhkan, keduanya
harus dilebarkan bersamaan.

Sales Order baru lahir berstatus 'waiting', dan itu persis keadaan yang
ditunggu halaman Draft Tally. Pemegang create_tallies kini diberi tahu saat
SO dibuat, hanya saat dibuat dan tidak saat disunting: menyunting SO yang
sama beberapa kali tidak melahirkan pekerjaan baru, dan notifikasi berulang
hanya membuat orang berhenti membacanya. Pembuatnya sendiri dikecualikan.

Sudah diperiksa dan ternyata aman: EditSalesOrder menghapus lalu membuat
ulang seluruh barisnya tiap kali menyimpan, sehingga ID barisnya berganti.
Tidak ada tabel mana pun yang menyimpan sales_order_item_id -- Invoice
mencocokkan lewat sales_order_id dan product_id.

Closes #165

+++ app/Models/Customer.php
@@ -25,7 +25,7 @@ class Customer extends Model
     ];
 
     protected $casts = [
-        'default_discount' => 'float',
+        'default_discount' => 'integer',
         'required_documents' => 'array',
         'invoice_exchange' => 'boolean',
         'is_taxable' => 'boolean',
